split spark bindings into several modules, to allow them to grow without becoming too annoying to work with
I also used this opportunity to add a few Spark methods here and there.
