+++ sparkle/src/Control/Distributed/Spark/ML/LDA.hs
@@ -1,0 +1,58 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE OverloadedStrings #-}
+module Control.Distributed.Spark.ML.LDA where
+
+import Control.Distributed.Spark.ML.Feature.CountVectorizer
+import Control.Distributed.Spark.PairRDD
+import Data.Coerce
+import Data.Int
+import Foreign.C.Types
+import Foreign.JNI
+
+newtype LDA = LDA (J ('Class "org.apache.spark.mllib.clustering.LDA"))
+
+newLDA :: Double                               -- ^ fraction of documents
+       -> Int32                                -- ^ number of topics
+       -> Int32                                -- ^ maximum number of iterations
+       -> IO LDA
+newLDA frac numTopics maxIterations = do
+  cls <- findClass "org/apache/spark/mllib/clustering/LDA"
+  lda <- newObject cls "()V" []
+
+  opti_cls <- findClass "org/apache/spark/mllib/clustering/OnlineLDAOptimizer"
+  opti <- newObject opti_cls "()V" []
+  setMiniBatch <- getMethodID opti_cls "setMiniBatchFraction" "(D)Lorg/apache/spark/mllib/clustering/OnlineLDAOptimizer;"
+  opti' <- callObjectMethod opti setMiniBatch [JDouble frac]
+
+  setOpti <- getMethodID cls "setOptimizer" "(Lorg/apache/spark/mllib/clustering/LDAOptimizer;)Lorg/apache/spark/mllib/clustering/LDA;"
+  lda' <- callObjectMethod lda setOpti [JObject opti']
+
+  setK <- getMethodID cls "setK" "(I)Lorg/apache/spark/mllib/clustering/LDA;"
+  lda'' <- callObjectMethod lda' setK [JInt numTopics]
+
+  setMaxIter <- getMethodID cls "setMaxIterations" "(I)Lorg/apache/spark/mllib/clustering/LDA;"
+  lda''' <- callObjectMethod lda'' setMaxIter [JInt maxIterations]
+
+  setDocConc <- getMethodID cls "setDocConcentration" "(D)Lorg/apache/spark/mllib/clustering/LDA;"
+  lda'''' <- callObjectMethod lda''' setDocConc [JDouble $ negate 1]
+
+  setTopicConc <- getMethodID cls "setTopicConcentration" "(D)Lorg/apache/spark/mllib/clustering/LDA;"
+  lda''''' <- callObjectMethod lda'''' setTopicConc [JDouble $ negate 1]
+
+  return (coerce (unsafeCast lda'''''))
+
+newtype LDAModel = LDAModel (J ('Class "org.apache.spark.mllib.clustering.LDAModel"))
+
+runLDA :: LDA -> PairRDD CLong SparkVector -> IO LDAModel
+runLDA lda rdd = do
+  cls <- findClass "Helper"
+  run <- getStaticMethodID cls "runLDA" "(Lorg/apache/spark/mllib/clustering/LDA;Lorg/apache/spark/api/java/JavaPairRDD;)Lorg/apache/spark/mllib/clustering/LDAModel;"
+  coerce . unsafeCast <$>
+    callStaticObjectMethod cls run [JObject lda, JObject rdd]
+
+describeResults :: LDAModel -> CountVectorizerModel -> Int32 -> IO ()
+describeResults lm cvm maxTerms = do
+  cls <- findClass "Helper"
+  mth <- getStaticMethodID cls "describeResults" "(Lorg/apache/spark/mllib/clustering/LDAModel;Lorg/apache/spark/ml/feature/CountVectorizerModel;I)V"
+  callStaticVoidMethod cls mth [JObject lm, JObject cvm, JInt maxTerms]
